Factor out AttemptsTable
