+++ docs/assets/js/components/ConfigTable.js
